Update GamifyServiceProvider.php
Make the look up of Gamifying badges be directory recursive so that it finds them no matter how deep the folder structure is. This allows developers to place badges within the `Gamify/Badges` directory in a more structured way rather than all badges right in the root of that directory.